Create separate file for files' methods
